cleanup + dir/plane included in game structure

+++ ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: gadeneux <gadeneux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/15 15:06:04 by gadeneux          #+#    #+#             */
-/*   Updated: 2022/05/30 19:53:19 by gadeneux         ###   ########.fr       */
+/*   Updated: 2022/06/08 15:33:09 by gadeneux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "header.h"
 
-int	ft_strncmp(char *s1, char *s2, int n)
+int	ft_strncmp(char *s1, char *s2, unsigned long long n)
 {
 	unsigned long long	i;
